getting back up to date

+++ scripts/util/counts_to_table2.py
@@ -173,7 +173,7 @@ for key in sorted(cD):
 for key in sorted(geneD):
     val = geneD.get(key)
     outGeneTable.write(key + '\t' + '\t'.join(map(str,val)) + '\n')
-    if all(item >= int(opts.threshold) for item in map(float,val)): #[1:])): # [1:] to take off the gene name
+    if all(item >= float(opts.threshold) for item in map(float,val)): #[1:])): # [1:] to take off the gene name
         outGeneThreshTable.write(key + '\t' + '\t'.join(map(str,val)) + '\n') #[1:])) + '\n') #take off gene name
 
 outCountTable.close()
